Gentle decay (power-law) mapping function from sizes to arrival times

Gentle decay (power-law) mapping function from sizes to arrival times

Usage

power_law_mapping(sizes, delta)

Arguments

sizes

vector of sizes to be mapped to arrival times.

delta

numeric value (greater than 0) for the exponent in the mapping function.

Value

vector of arrival times.

Examples

# Generate a vector of sizes
sizes <- runif(100, 0, 100)

# Map the sizes to arrival times using a power-law mapping with delta = .5
power_law_mapping(sizes, .5)
#>   [1] 0.25996293 0.02284869 0.02582357 0.25292089 0.63257739 0.50083185
#>   [7] 0.44051842 0.01711844 0.00000000 0.09126123 0.17561348 0.58563945
#>  [13] 0.08481810 0.02048083 0.23972569 0.18498574 0.25088306 0.32256700
#>  [19] 0.62060506 0.05314732 0.82248364 0.50945225 0.70412833 0.32017716
#>  [25] 0.54723892 0.20851635 0.88167074 0.49517041 0.03966295 0.24509560
#>  [31] 0.00966031 0.73429491 0.84652148 0.01850379 0.29310817 0.68404290
#>  [37] 0.07144187 0.61388022 0.47277926 0.87265838 0.74537820 0.23581020
#>  [43] 0.69905756 0.16925240 0.42355140 0.03091506 0.28752230 1.00000000
#>  [49] 0.04690740 0.11497317 0.87636968 0.77231154 0.09140686 0.08011394
#>  [55] 0.27078485 0.09969468 0.39891642 0.26114027 0.46748972 0.39375624
#>  [61] 0.38559160 0.02901611 0.19662302 0.19982140 0.54595745 0.32961402
#>  [67] 0.25800174 0.19709525 0.58603227 0.33198244 0.33517588 0.80660724
#>  [73] 0.65076381 0.33833723 0.72208856 0.50634109 0.19870653 0.08926727
#>  [79] 0.10495270 0.18730027 0.43215439 0.29583933 0.27238206 0.11733587
#>  [85] 0.65621884 0.04202949 0.25961789 0.07181502 0.26226508 0.16823526
#>  [91] 0.65869241 0.13067081 0.19734849 0.49922103 0.24698847 0.45130674
#>  [97] 0.52391384 0.18318604 0.13408940 0.03807729